modularised structure with lib.rs added
